Pull out traversal functions into separate module and constituent functions.
